Renamed common catalog to shared
