+++ src/app/modules/courses/courses.component.ts
@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 // import { CommonModule } from '@angular/common';
 // import { CoursesListItemComponent } from './courses-list-item/courses-list-item.component';
-import { ICourse } from 'src/app/common/interfaces/course';
+import { ICourse } from 'src/app/shared/interfaces/course';
 
 @Component({
   selector: 'app-courses',
