+++ src/app/shared/interfaces/course.ts
